@@ -1,7 +1,8 @@
 /* eslint-disable import/namespace */
 import Mustache from 'mustache'
 import * as langs from './langs/index.mjs'
-import { DEFAULT_LANGUAGE } from '@juquinha/config/env.mjs'
+
+const DEFAULT_LANGUAGE = 'en'
 
 export default (term, params, { language } = {}) => {
   const lang = (language || DEFAULT_LANGUAGE || '').replace('-', '')
