@@ -1,121 +1,18 @@
-import colors from 'vuetify/es5/util/colors'
-import * as config from '@juquinha/config/env.mjs'
-import { defineNuxtConfig } from '@nuxt/bridge'
+import path from 'path'
+import { defineNuxtConfig } from 'nuxt'
 
+// https://v3.nuxtjs.org/api/configuration/nuxt.config
 export default defineNuxtConfig({
-  // Disable server-side rendering: https://go.nuxtjs.dev/ssr-mode
-  ssr: false,
-
-  env: {
-    ...config,
-  },
-
-  // Target: https://go.nuxtjs.dev/config-target
-  target: 'static',
-
-  // Global page headers: https://go.nuxtjs.dev/config-head
-  head: {
-    titleTemplate: '%s - web',
-    title: 'web',
-    meta: [
-      { charset: 'utf-8' },
-      { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      { hid: 'description', name: 'description', content: '' },
-      { name: 'format-detection', content: 'telephone=no' }
-    ],
-    link: [
-      { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }
-    ]
-  },
-
-  // Global CSS: https://go.nuxtjs.dev/config-css
-  css: [
-  ],
-
-  // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
-  plugins: [
-    '~/plugins/i18n',
-    '~/plugins/api',
-    '~/plugins/noty',
-    '~/plugins/vuelidate',
-    '~/plugins/vue-the-mask',
-  ],
-
-  // Auto import components: https://go.nuxtjs.dev/config-components
-  components: true,
-
-  // Modules for dev and build (recommended): https://go.nuxtjs.dev/config-modules
-  buildModules: [
-    // https://go.nuxtjs.dev/stylelint
-    '@nuxtjs/stylelint-module',
-    // https://go.nuxtjs.dev/vuetify
-    '@nuxtjs/vuetify',
-  ],
-
-  // Modules: https://go.nuxtjs.dev/config-modules
-  modules: [
-    // https://go.nuxtjs.dev/axios
-    '@nuxtjs/axios',
-    // https://go.nuxtjs.dev/pwa
-    '@nuxtjs/pwa',
-  ],
-
-  // Axios module configuration: https://go.nuxtjs.dev/config-axios
-  axios: {
-    baseURL: config.API_URL || '',
-    paramsSerializer: function paramsSerializer(params) {
-      return qs.stringify(params, { arrayFormat: 'brackets' })
-    },
-    retry: { retries: 1 },
-  },
-
-  // PWA module configuration: https://go.nuxtjs.dev/pwa
-  pwa: {
-    manifest: {
-      lang: 'en'
-    }
-  },
-
-  // Vuetify module configuration: https://go.nuxtjs.dev/config-vuetify
-  vuetify: {
-    customVariables: ['~/assets/variables.scss'],
-    theme: {
-      dark: false,
-      themes: {
-        light: {
-          primary: colors.lightBlue.darken1,
-          accent: colors.lightBlue.darken4,
-          secondary: colors.lightGreen.darken1,
-          info: colors.lightBlue.darken2,
-          warning: colors.amber.darken1,
-          error: colors.deepOrange.darken1,
-          success: colors.lightGreen.darken3,
-        },
-        // dark: {
-        //   primary: colors.blue.darken2,
-        //   accent: colors.grey.darken3,
-        //   secondary: colors.amber.darken3,
-        //   info: colors.teal.lighten1,
-        //   warning: colors.amber.base,
-        //   error: colors.deepOrange.accent4,
-        //   success: colors.green.accent3
-        // }
-      }
-    }
-  },
-
-  // Build Configuration: https://go.nuxtjs.dev/config-build
   build: {
-    // @see https://github.com/nuxt-community/dotenv-module/issues/11
-    extend(config) {
-      config.node = {
-        fs: 'empty',
-      }
-    },
+    transpile: ['vuetify'],
   },
-
-  // See https://github.com/nuxt/framework/issues/886
-  bridge: {
-    nitro: false
+  vuetify: {
+    customVariables: true,
+    treeShake: true
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'),
+    },
   },
 })
